+++ sources/Valkey.Glide/Internals/Cmd.cs
@@ -109,6 +109,12 @@ internal class Cmd<R, T> : ICmd
         RequestType.PubSubShardChannels => ["PUBSUB", "SHARDCHANNELS"],
         RequestType.PubSubShardNumSub => ["PUBSUB", "SHARDNUMSUB"],
         RequestType.GetSubscriptions => ["GET_SUBSCRIPTIONS"],
+        RequestType.FtList => ["FT._LIST"],
+        RequestType.FtAggregate => ["FT.AGGREGATE"],
+        RequestType.FtCreate => ["FT.CREATE"],
+        RequestType.FtDropIndex => ["FT.DROPINDEX"],
+        RequestType.FtInfo => ["FT.INFO"],
+        RequestType.FtSearch => ["FT.SEARCH"],
         _ => [requestType.ToString().ToUpper()]
     };
 #pragma warning restore IDE0072 // Populate switch