+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 - Add support for Windows CI and testing with WSL (#184)
 - Add StackExchange.Redis compatible pub/sub API (#202)
 - Add transparent compression support with Zstd and LZ4 backends (#213)
+- Add Valkey Search command support (#225)